backend/requirements.txt
django==3.2.25
djongo==1.3.6
pymongo==3.12.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
